environment fixes
