+++ config/envs/authoring.yaml
@@ -43,24 +43,44 @@ servers:
         spec:
           replicas: 1
           template:
+            metadata:
+              name: "{{.ApplicationName}}-kieserver-{{$index}}"
             spec:
               containers:
                 - name: "{{.ApplicationName}}-kieserver-{{$index}}"
                   env:
                     - name: MAVEN_REPOS
                       value: "RHPAMCENTR,EXTERNAL"
-                    - name: RHPAMCENTR_MAVEN_REPO_USERNAME
-                      value: mavenUser
-                    - name: RHPAMCENTR_MAVEN_REPO_PASSWORD
-                      value: "{{.MavenPassword}}"
-                    - name: RHPAMCENTR_MAVEN_REPO_PATH
-                      value: /maven2/
-                    - name: RHPAMCENTR_MAVEN_REPO_SERVICE
-                      value: "{{.ApplicationName}}-rhpamcentr"
                     - name: KIE_SERVER_CONTROLLER_TOKEN
                       value: ""
-                    - name: KIE_SERVER_ID
-                      value: "{{.ApplicationName}}-kieserver-{{$index}}"
+                    - name: DATASOURCES
+                      value: "RHPAM"
+                    - name: RHPAM_DATABASE
+                      value: "rhpam7"
+                    - name: RHPAM_JNDI
+                      value: "java:/jboss/datasources/rhpam"
+                    - name: RHPAM_JTA
+                      value: "true"
+                    - name: KIE_SERVER_PERSISTENCE_DS
+                      value: "java:/jboss/datasources/rhpam"
+                    ## H2 driver settings BEGIN
+                    - name: RHPAM_DRIVER
+                      value: "h2"
+                    - name: RHPAM_USERNAME
+                      value: "sa"
+                    - name: RHPAM_PASSWORD
+                      value: "{{.AdminPassword}}"
+                    - name: RHPAM_SERVICE_HOST
+                      value: "dummy_ignored"
+                    - name: RHPAM_SERVICE_PORT
+                      value: "12345"
+                    - name: KIE_SERVER_PERSISTENCE_DIALECT
+                      value: "org.hibernate.dialect.H2Dialect"
+                    - name: DROOLS_SERVER_FILTER_CLASSES
+                      value: "true"
+                    - name: RHPAM_XA_CONNECTION_PROPERTY_URL
+                      value: "jdbc:h2:/opt/eap/standalone/data/rhpam"
+                  ## H2 driver settings END
                   volumeMounts:
                     - name: "{{.ApplicationName}}-h2-pvol"
                       mountPath: "/opt/eap/standalone/data"
@@ -69,14 +89,21 @@ servers:
                 - name: "{{.ApplicationName}}-h2-pvol"
                   persistentVolumeClaim:
                     claimName: "{{.ApplicationName}}-h2-claim-{{$index}}"
-<<<<<<< HEAD
-  ## KIE server deployment config END
-  #{{end}}
-  ## RANGE ends
-  ## KIE Servers END
-=======
     ## KIE server deployment config END
+    persistentVolumeClaims:
+      ## H2 persistent volume claim BEGIN
+      - metadata:
+          name: "{{.ApplicationName}}-h2-claim-{{$index}}"
+          labels:
+            app: "{{.ApplicationName}}"
+            service: "{{.ApplicationName}}-kieserver-{{$index}}"
+        spec:
+          accessModes:
+            - ReadWriteOnce
+          resources:
+            requests:
+              storage: "1Gi"
+      ## H2 persistent volume claim END
 #{{end}}
 ## RANGE ends
 ## KIE Servers END
->>>>>>> upstream/master
